+++ examples/simple/index.js
@@ -2,30 +2,26 @@ import 'dotenv/config'
 import { Content, Email, EmailAddress, Templateless } from 'templateless-js'
 
 const sendEmail = async () => {
-  const api_key = process.env.TEMPLATELESS_API_KEY
-  if (!api_key) {
+  const apiKey = process.env.TEMPLATELESS_API_KEY
+  if (!apiKey) {
     console.error('Set TEMPLATELESS_API_KEY to your Templateless API key')
     return
   }
 
-  const email_address = process.env.TEMPLATELESS_EMAIL_ADDRESS
-  if (!email_address) {
+  const emailAddress = process.env.TEMPLATELESS_EMAIL_ADDRESS
+  if (!emailAddress) {
     console.error('Set TEMPLATELESS_EMAIL_ADDRESS to your own email address')
     return
   }
 
-  const content = Content.builder().text('Hello world').build()
-
   const email = Email.builder()
-    .to(new EmailAddress(email_address))
-    .subject('Hello')
-    .content(content)
+    .to(new EmailAddress(emailAddress))
+    .subject('Hello 👋')
+    .content(Content.builder().text('Hello world').build())
     .build()
 
-  const templateless = new Templateless(api_key)
-  const result = await templateless.send(email)
-
-  console.log(result)
+  const templateless = new Templateless(apiKey)
+  await templateless.send(email)
 }
 
 sendEmail()