+++ src/common.ts
@@ -7,4 +7,9 @@ export enum ComponentId {
   Socials = 'SOCIALS',
   Text = 'TEXT',
   ViewInBrowser = 'VIEW_IN_BROWSER',
+  StoreBadges = 'STORE_BADGES',
+  QrCode = 'QR_CODE',
+  Signature = 'SIGNATURE',
 }
+
+export interface Component {}