+++ src/components/index.ts
@@ -5,6 +5,9 @@ import { Otp } from './otp'
 import { Service, Item as SocialItem, Socials } from './socials'
 import { Text } from './text'
 import { ViewInBrowser } from './viewInBrowser'
+import { StoreBadge, Item as StoreBadgeItem, StoreBadges } from './storeBadges'
+import { QrCode } from './qrCode'
+import { Signature, Font as SignatureFont } from './signature'
 
 export { Button }
 export { Image }
@@ -13,6 +16,9 @@ export { Otp }
 export { Service, SocialItem, Socials }
 export { Text }
 export { ViewInBrowser }
+export { StoreBadges, StoreBadgeItem, StoreBadge }
+export { QrCode }
+export { Signature, SignatureFont }
 
 export type Component =
   | Button
@@ -24,3 +30,6 @@ export type Component =
   | Socials
   | Text
   | ViewInBrowser
+  | StoreBadges
+  | QrCode
+  | Signature