+++ src/content.ts
@@ -8,6 +8,11 @@ import {
   Text,
   ViewInBrowser,
   SocialItem,
+  QrCode,
+  StoreBadges,
+  StoreBadgeItem,
+  Signature,
+  SignatureFont,
 } from './components'
 import { Header, Footer } from '.'
 
@@ -96,6 +101,22 @@ class ContentBuilder {
 
   viewInBrowser(text?: string): this {
     return this.push(new ViewInBrowser(text))
+  }
+
+  qrCode(url: string): this {
+    return this.push(QrCode.url(url))
+  }
+
+  storeBadges(data: StoreBadgeItem[]): this {
+    return this.push(new StoreBadges(data))
+  }
+
+  signature(text: string, font?: SignatureFont): this {
+    return this.push(new Signature(text, font))
+  }
+
+  component(c: Component): this {
+    return this.push(c)
   }
 
   build(): Content {