+++ CHANGELOG.md
@@ -1,17 +1,30 @@
 # CHANGELOG
 
-**v0.1.0-alpha.4:**
+## v0.1.0-alpha.5:
+
+### New Features
+- New social icons: `Service.Mastodon` and `Service.Rss`
+- New `StoreBadges` component
+- New `QrCode` component
+- New `Signature` component
+- New [examples](examples)
+
+### Enhancements
+- Updated README
+- Dependency updates
+
+## v0.1.0-alpha.4:
 - `README.md`: notice about test mode
 - Support for test mode logging
 - npm updates
 
-**v0.1.0-alpha.3:**
+## v0.1.0-alpha.3:
 - `README.md` cleanup (listing of components)
 - `Image` component now requires only `src`; the other params are optional
 - `ViewInBrowser` component has changed: text is optional
 - npm updates
 
-**v0.1.0-alpha.2:**
+## v0.1.0-alpha.2:
 - Introduced `CHANGELOG.md`
 - npm updates
 - Introduced new services as [social icons](examples/confirm_email/index.js):
@@ -29,5 +42,5 @@
 - **Breaking Changes:**
   - Renamed `Service.Github` to `Service.GitHub`
 
-**v0.1.0-alpha.1:**
+## v0.1.0-alpha.1:
 - Initial implementation